Refactor tests to share between sessions and metrics Make queries work using metrics dataset
